@@ -1,11 +1,11 @@
 {
   "name": "vuse-rx",
-  "version": "0.10.0",
+  "version": "0.12.0-rc.17",
   "sideEffects": false,
   "description": "First-class rxjs support for Vue 3",
   "main": "dist/js",
   "module": "dist/es",
-  "unpkg": "dist/esm",
+  "unpkg": "dist/umd",
   "typings": "src",
   "files": [
     "dist",
@@ -25,12 +25,12 @@
     "checks": "madge --circular --extensions ts src/index.ts",
     "commit-build": "(git diff --quiet && git diff --staged --quiet) || (git commit -am \"Update dist\")",
     "preversion": "npm t && npm run build && npm run commit-build",
-    "prerelease": "npm version prerelease --preid=rc && npm publish --tag next && git push",
-    "pre-minor": "npm version preminor --preid=rc && npm publish --tag next && git push",
-    "pre-major": "npm version premajor --preid=rc && npm publish --tag next && git push",
-    "patch": "npm version patch && npm publish && git push",
-    "minor": "npm version minor && npm publish && git push",
-    "major": "npm version major && npm publish && git push",
+    "prerelease": "npm version prerelease --preid=rc",
+    "pre-minor": "npm version preminor --preid=rc",
+    "pre-major": "npm version premajor --preid=rc",
+    "patch": "npm version patch",
+    "minor": "npm version minor",
+    "major": "npm version major",
     "docs:dev": "vitepress dev docs",
     "docs:build": "vitepress build docs",
     "docs:serve": "vitepress serve docs"
@@ -41,6 +41,8 @@
     "rxjs",
     "vue3",
     "rx",
+    "simple",
+    "pure",
     "integration",
     "support",
     "composition-api",
@@ -74,11 +76,11 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.0",
     "jest": "^26.6.3",
-    "madge": "^4.0.0",
+    "madge": "^4.0.2",
     "rollup-plugin-typescript2": "^0.29.0",
     "rxjs": "^6.6.3",
     "ts-jest": "^26.5.0",
-    "typescript": "^4.1.3",
+    "typescript": "^4.2.3",
     "vitepress": "^0.12.2",
     "vue": "^3.0.5",
     "vuse-rx": "./"
