@@ -1,6 +1,6 @@
-import { BehaviorSubject, isObservable, noop, of } from 'rxjs';
-import { Observable } from 'rxjs/internal/Observable';
-import { delay, tap } from 'rxjs/operators';
+import type { Observable } from 'rxjs';
+import { from, isObservable, noop } from 'rxjs';
+import { tap } from 'rxjs/operators';
 import { isProxy } from 'vue';
 import { useRxState } from '../src/use-rx';
 
@@ -11,22 +11,21 @@ const fn = jest.fn();
 const defaultCount = 0;
 const step = 1;
 
+const resolveCount = (count: number = 0) => Promise.resolve({ count });
+
 const use = () => useRxState<State>(() => ({ count: defaultCount }))({
   increment: () => state => ({ count: state.count + step }),
   setCount: (count: number) => ({ count }),
-  setCountAfter: (count: number, timeout: number) => of({ count }).pipe(delay(timeout)),
-  incrementAfter(timeout: number) {
+  setCountAsync: (count: number) => from(resolveCount(count)),
+  incrementAsync() {
     return (state): Observable<State> =>
-      this.setCountAfter(this.increment()(state).count, timeout);
+      this.setCountAsync(this.increment()(state).count);
   },
   setFoo: (foo?: string) => ({ foo }),
 }, state$ => state$.pipe(tap(fn)));
 
-const time = (timeout: number) => new Promise(res => setTimeout(() => res(undefined), timeout));
-
 describe('useRxState', () => {
   it('binds the state to reducers', async () => {
-    const timeout = 100;
     const $subscription = jest.fn();
 
     const { actions, actions$, state, state$ } = use().subscribe($subscription);
@@ -40,23 +39,23 @@ describe('useRxState', () => {
     expect(state.count).toBe(defaultCount);
     expect(state.foo).toBeUndefined();
 
-    expect(actions.incrementAfter(timeout)).toBeUndefined();
+    expect(actions.incrementAsync()).toBeUndefined();
 
     expect(state.count).toBe(defaultCount);
 
-    await time(timeout);
+    await resolveCount();
 
     expect(fn).toHaveBeenCalled();
 
     expect(state.count).toBe(defaultCount + step);
     expect($subscription).toBeCalledTimes(1);
 
-    expect(actions.setCountAfter(defaultCount, timeout)).toBeUndefined();
+    expect(actions.setCountAsync(defaultCount)).toBeUndefined();
 
     expect(state.count).toBe(defaultCount + step);
     expect($subscription).toBeCalledTimes(1);
 
-    await time(timeout);
+    await resolveCount();
 
     expect(state.count).toBe(defaultCount);
     expect($subscription).toBeCalledTimes(2);
@@ -151,7 +150,7 @@ describe('useRxState', () => {
     const wrong = useRxState({ value: 0 })({
       setValue: (value: number) => (_, ctx) => {
         if (value < 0) {
-          ctx.error(error);
+          ctx?.error(error);
 
           // signify that no mutation happens
           return {};
@@ -170,7 +169,7 @@ describe('useRxState', () => {
     const right = useRxState({ value: 0 })({
       setValue: (value: number) => (_, ctx) => {
         if (value === 42) {
-          ctx.complete();
+          ctx?.complete();
         }
 
         return { value };
