@@ -1,4 +1,4 @@
 export * from './use-rx';
-export * from './rx-refs/index';
+export * from './operators';
+export * from './rx-refs';
 export * from './hooks/from';
-export * from './hooks/until';
