@@ -1,5 +1,6 @@
 import { from, ObservableInput } from 'rxjs';
 import { isProxy, Ref, ref, toRef, UnwrapRef } from 'vue';
+import { untilUnmounted } from '../operators/until';
 
 /**
  * Creates a ref from a promise.
@@ -22,6 +23,15 @@ export function refFrom<R>(promise: Promise<R>): Ref<UnwrapRef<R> | undefined>;
  * @param defaultValue to set to the ref initially
  */
 export function refFrom<R>(promise: Promise<R>, defaultValue: R): Ref<UnwrapRef<R>>;
+
+/**
+ * `refFrom` is not supposed to work for dynamic types and is recommended to use statically!
+ *
+ * But this overload is still allowed as a convenience.
+ *
+ * @param value a value to create the ref from
+ */
+export function refFrom<R>(arg: R): Ref<UnwrapRef<R>>;
 
 /**
  * Creates a ref from an observable input.
@@ -81,7 +91,9 @@ export function refFrom(arg: unknown, subArg?: unknown) {
   if (typeof arg === 'object') try {
     const ref$ = ref(subArg);
 
-    from(arg as any).subscribe(value => ref$.value = value);
+    untilUnmounted(from(arg as any)).subscribe({
+      next: value => ref$.value = value
+    });
 
     return ref$;
   } catch (_) { /* Silence the error to try another ways */ }
