@@ -1,10 +1,14 @@
-import { ObservableInput, from } from 'rxjs';
+import { ObservableInput, from, Subscription, Observable } from 'rxjs';
 import { Ref, ref } from 'vue';
+import { untilUnmounted } from '../operators/until';
 
 type Subscribers<R, E> = { next: R; error: E; };
 
-type Refs<S extends Subscribers<unknown, unknown>> = {
-  [key in keyof S]: Ref<S[key]>;
+type Refs<R, E> = {
+  [key in keyof Subscribers<R, E>]: Ref<Subscribers<R, E>[key]>;
+} & {
+  subscription: Subscription;
+  value$: Observable<R>;
 };
 
 /**
@@ -18,7 +22,7 @@ type Refs<S extends Subscribers<unknown, unknown>> = {
  * A value to create an observable from.\
  * This observable is then going to be listened to in order to set the refs.
  */
-export function refsFrom<R, E = unknown>(input: ObservableInput<R>): Refs<Subscribers<R | undefined, E | undefined>>;
+export function refsFrom<R, E = unknown>(input: ObservableInput<R>): Refs<R | undefined, E | undefined>;
 
 /**
  * Creates two refs from an observable input (promise, iterable, observable and alike):
@@ -32,7 +36,7 @@ export function refsFrom<R, E = unknown>(input: ObservableInput<R>): Refs<Subscr
  * @param defaultValues
  * A map of default values for each ref.
  */
-export function refsFrom<R, E = unknown>(input: ObservableInput<R>, defaultValues: { next: R }): Refs<Subscribers<R, E | undefined>>;
+export function refsFrom<R, E = unknown>(input: ObservableInput<R>, defaultValues: { next: R }): Refs<R, E | undefined>;
 
 /**
  * Creates two refs from an observable input (promise, iterable, observable and alike):
@@ -46,16 +50,20 @@ export function refsFrom<R, E = unknown>(input: ObservableInput<R>, defaultValue
  * @param defaultValues
  * A map of default values for each ref.
  */
-export function refsFrom<R, E>(input: ObservableInput<R>, defaultValues: Subscribers<R, E>): Refs<Subscribers<R, E>>;
+export function refsFrom<R, E>(input: ObservableInput<R>, defaultValues: Subscribers<R, E>): Refs<R, E>;
 
-export function refsFrom(input: ObservableInput<any>, defaultValues?: Partial<Subscribers<any, any>>) {
-  const next = ref(defaultValues?.next);
-  const error = ref(defaultValues?.error);
+export function refsFrom(input: ObservableInput<any>, defaultValues: Partial<Subscribers<any, any>> = {}) {
+  const next = ref(defaultValues.next);
+  const error = ref(defaultValues.error);
+  const value$ = untilUnmounted(from(input));
 
-  from(input).subscribe({
-    next: v => next.value = v,
-    error: v => error.value = v,
-  });
-
-  return { next, error };
+  return {
+    next,
+    error,
+    value$,
+    subscription: value$.subscribe({
+      next: v => next.value = v,
+      error: v => error.value = v,
+    })
+  };
 }
