@@ -1,7 +1,7 @@
 import { isObservable, merge, of, Subject } from 'rxjs';
 import { map, mergeScan, scan, tap } from 'rxjs/operators';
 import { reactive, readonly } from 'vue';
-import { untilUnmounted } from './hooks/until';
+import { untilUnmounted } from './operators/until';
 
 import type { Observable, PartialObserver, Subscription } from 'rxjs';
 import type { DeepReadonly, Ref, UnwrapRef } from 'vue';
@@ -18,37 +18,9 @@ export interface MutationStrategy<S extends Record<PropertyKey, any>> {
   ) => S;
 }
 
-export const canMergeDeep = <S extends Record<PropertyKey, any>>(
-  state: S,
-  mutation: S | DeepPartial<S>,
-  key: keyof S,
-) =>  (
-  typeof mutation[key] === 'object'
-  && mutation !== null
-  && typeof state[key] === 'object'
-);
-
-export const deepMergeKeys = <S extends Record<PropertyKey, any>>(
-  state: S
-) => (
-  mutation: DeepPartial<S>
-) => {
-  for (const key in mutation) {
-    state[key] = canMergeDeep(state, mutation, key)
-      ? deepMergeKeys(state[key])(mutation[key])
-      : mutation[key] as any;
-  }
-
-  return state;
-};
-
-export interface RxStateOptions<S> {
+export interface RxStateOptions<S extends Record<PropertyKey, any>> {
   mutationStrategy: MutationStrategy<S>;
 }
-
-const defaultOptions: RxStateOptions<any> = {
-  mutationStrategy: deepMergeKeys,
-};
 
 /**
  * Allows to bind reducers to a state and an observable.
@@ -59,16 +31,18 @@ const defaultOptions: RxStateOptions<any> = {
  * @param initialState - a factory or initial value for the reactive state
  * @param options - options to customize the behavior, for example - to apply a custom strategy of merging a mutation with an old state
  */
-export function useRxState<T extends Record<string, any>>(
+export function useRxState<T extends Record<PropertyKey, any>>(
   initialState: T | (() => T),
   options: Partial<RxStateOptions<UnwrapNestedRefs<T>>> = defaultOptions
 ): CreateRxState<UnwrapNestedRefs<T>> {
-  const {
-    mutationStrategy: mergeKeys,
-  } = { ...defaultOptions, ...options };
+  type S = UnwrapNestedRefs<T>;
+
+  const { mutationStrategy: mergeKeys } = {
+    ...defaultOptions as RxStateOptions<S>,
+    ...options
+  };
 
   return function (reducers, map$?) {
-    type S = UnwrapNestedRefs<T>;
     type ReducerResult = ReturnType<StateReducer<S>>;
     type Actions = ReducerActions<typeof reducers>;
 
@@ -78,11 +52,19 @@ export function useRxState<T extends Record<string, any>>(
     const actions$ = <ReducerObservables<Actions, S>> {};
     const actions$Arr = <Observable<S>[]> [];
 
+    let complete = false;
+    let error: any = undefined;
+
+    const context: MutationContext = {
+      error: e => { error = e; },
+      complete: () => { complete = true; }
+    };
+
     for (const key in reducers) {
       const mutations$ = new Subject<ReducerResult>();
 
       actions[key] = <ReducerAction<typeof reducers[typeof key]>>(
-        (...args) => mutations$.next(
+        (...args: any[]) => mutations$.next(
           reducers[key].apply(reducers, args)
         )
       );
@@ -90,23 +72,19 @@ export function useRxState<T extends Record<string, any>>(
       actions$Arr.push(
         actions$[`${key}$` as const] = (
           mergeScan((prev: S, curr: ReducerResult) => {
-            let complete = false;
-            let error: any = undefined;
-
-            curr = maybeCall(curr, prev, {
-              error: e => { error = e },
-              complete: () => { complete = true }
-            });
+            curr = maybeCall(curr, prev, context);
 
             return (
-              isObservable(curr) ? curr : of(curr)
+              isObservable(curr)
+                ? curr
+                : of(curr)
             ).pipe(
               map(mergeKeys(prev, mergeKeys)),
-              tap(() => (
-                complete
-                  ? mutations$.complete()
-                  : error && mutations$.error(error)
-              ))
+              tap({
+                next: () => error
+                  ? error = mutations$.error(error)
+                  : complete && mutations$.complete()
+              })
             )
           }, state)(mutations$)
         )
@@ -124,6 +102,7 @@ export function useRxState<T extends Record<string, any>>(
           reducers,
           state,
           actions$,
+          context,
         ).pipe(
           scan((prev, curr) => (mergeKeys(prev, mergeKeys)(curr)), state)
         ) : merged$
@@ -155,8 +134,8 @@ type CreateRxState<S> = {
    *   })
    *   ```
    *
-   * @param initialState an initial value for the reactive state
-   * @param mergeStrategy a strategy of merging a mutation with an old state
+   * @param reducers a map of reducers to mutate the state
+   * @param map$ a function to modify the resulting observable
    */
   <R extends StateReducers<S>>(
     reducers: R,
@@ -164,9 +143,54 @@ type CreateRxState<S> = {
       state$: Observable<S>,
       reducers: R,
       state: S,
-      actions$: Record<Action$<Extract<keyof R, string>>, Observable<S>>
+      actions$: Record<Action$<Extract<keyof R, string>>, Observable<S>>,
+      context: MutationContext
     ) => Observable<DeepPartial<S>>
-  ): SubscribableRxRes<ReducerActions<R>, S>;
+  ): SubscribableRxResult<ReducerActions<R>, S>;
+};
+
+/**
+ * Checks if it's possible to advance deeper
+ * into the sibling object structures,
+ * with one being partial
+ *
+ * @param state - the object source
+ * @param mutation - the main checking reference
+ * @param key - a key into which to advance
+ */
+export const canMergeDeep = <S extends Record<PropertyKey, any>>(
+  state: S,
+  mutation: S | DeepPartial<S>,
+  key: keyof S,
+) => (
+  typeof mutation[key] === 'object'
+  && mutation !== null
+  && typeof state[key] === 'object'
+);
+
+/**
+ * Default merge strategy for mutations
+ *
+ * Merges state and mutation recursively,
+ * by enumerable keys (`for..in`),
+ * so avoid recursive object links
+ */
+export const deepMergeKeys = <S extends Record<PropertyKey, any>>(
+  state: S
+) => (
+  mutation: S | DeepPartial<S>
+) => {
+  for (const key in mutation) {
+    state[key as keyof S] = canMergeDeep(state, mutation, key)
+      ? deepMergeKeys(state[key])(mutation[key])
+      : mutation[key] as any;
+  }
+
+  return state;
+};
+
+const defaultOptions = {
+  mutationStrategy: deepMergeKeys,
 };
 
 const createRxResult = <S, Actions>(result: {
@@ -174,9 +198,9 @@ const createRxResult = <S, Actions>(result: {
   state: DeepReadonly<S>,
   state$: Observable<S>,
   actions$: ReducerObservables<Actions, DeepReadonly<S>>
-}): SubscribableRxRes<Actions, S> => ({
+}): SubscribableRxResult<Actions, S> => ({
   ...result,
-  subscribe: (...args) => ({
+  subscribe: (...args: any[]) => ({
     ...result,
     subscription: result.state$.subscribe(...args),
   }),
@@ -217,26 +241,29 @@ export type DeepPartial<T> = T extends Builtin
 
 type UnwrapNestedRefs<T> = T extends Ref ? T : UnwrapRef<T>;
 
-export type MutationContext = {
+export interface MutationContext {
   error(error: any): void;
   complete(): void;
 }
 
-type ShallowReturn<S> =
+export type Mutation<S> =
   // | void
   // isn't useful,
-  // need to keep mutations explicit in what they do
+  // need to keep reducers explicit in what they do
   | S
   | DeepPartial<S>
   | Observable<DeepPartial<S>>;
 
+export type StatefulMutation<S> = (state: S, mutation?: MutationContext) => Mutation<S>;
+
 /**
  * A reducer for the observable state
  */
-export type StateReducer<S, Args extends any[] = any[]> = (...args: Args) =>
-  | ShallowReturn<S>
-  | ((state: S, mutation: MutationContext) => ShallowReturn<S>)
-  ;
+export type StateReducer<S, Args extends any[] = any[]> = (
+  (...args: Args) =>
+    | Mutation<S>
+    | StatefulMutation<S>
+);
 
 /**
  * A named collection of state reducers
@@ -248,6 +275,12 @@ export type StateReducers<S> = Record<string, StateReducer<S>>;
  */
 export type ResAction<A extends any[] = []> = (...args: A) => void;
 
+type Action$<Name extends string> = `${Name}$`;
+
+type ReducerObservables<H, R> = {
+  [key in Action$<Extract<keyof H, string>>]: Observable<R>;
+};
+
 /**
  * Resulting RX bindings:
  *
@@ -255,39 +288,23 @@ export type ResAction<A extends any[] = []> = (...args: A) => void;
  * * state - a reactive vue state
  * * state$ - an rxjs observable
  */
-export type RxResult<
-  Actions,
-  State,
-  RState = DeepReadonly<State>
-> = {
+export interface RxResult<Actions, State> {
   readonly actions: Actions;
-  readonly state: RState;
+  readonly state: DeepReadonly<State>;
   readonly state$: Observable<State>;
+  readonly actions$: ReducerObservables<Actions, DeepReadonly<State>>;
 };
 
-type Action$<Name extends string> = `${Name}$`;
-
-type ReducerObservables<H, R> = {
-  [key in Action$<Extract<keyof H, string>>]: Observable<R>;
-};
-
-export type SubscribableRxRes<
-  Actions,
-  State,
-  RState = DeepReadonly<State>
-> = RxResult<Actions, State, RState> & {
-  readonly actions$: ReducerObservables<Actions, RState>;
-  readonly subscribe: PipeSubscribe<SubscribableRxRes<Actions, State, RState>, State>;
-};
-
-export type PipeSubscribe<Res extends SubscribableRxRes<any, any>, S> = {
-  (observer?: PartialObserver<S>): Omit<Res, 'subscribe'> & {
-    readonly subscription: Subscription;
-  };
-  (...args: Parameters<Observable<S>['subscribe']>): Omit<Res, 'subscribe'> & {
-    readonly subscription: Subscription;
+export interface SubscribableRxResult<Actions, State> extends RxResult<Actions, State> {
+  readonly subscribe: {
+    (observer?: PartialObserver<State>): SubscriptionRxResult<Actions, State>;
+    (...args: Parameters<Observable<State>['subscribe']>): SubscriptionRxResult<Actions, State>;
   };
 };
+
+export interface SubscriptionRxResult<Actions, State> extends RxResult<Actions, State> {
+  readonly subscription: Subscription;
+}
 
 type ReducerAction<R> = R extends StateReducer<any, infer Args>
   ? ResAction<Args>
